+++ hera-backend/src/main/java/com/hera/backend/entity/ProductoNotaId.java
@@ -14,5 +14,5 @@ import java.io.Serializable;
 public class ProductoNotaId implements Serializable {
 
     private Long productoId;
-    private Integer notaId;
+    private Long notaId;
 }
