+++ hera-backend/src/main/java/com/hera/backend/repository/NivelDisponibilidadRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface NivelDisponibilidadRepository extends JpaRepository<NivelDisponibilidad, Integer> {
+public interface NivelDisponibilidadRepository extends JpaRepository<NivelDisponibilidad, Long> {
 
     Optional<NivelDisponibilidad> findByCodigo(String codigo);
     Optional<NivelDisponibilidad> findByNombre(String nombre);
